Update Client.java
Client allows user to register to the central server by using a socket and then thereby registering all its local files to the server. Client can also perform a search by sending a query to the central server to list all the peers containing that searched file. User can select the peer from where it wants to download the file.